Moved query for external wind to OnSensorUpdate and used Haeding from Larus
The previously used track is not available on the ground, giving the impression that the wind arrows are not responding. Furthermore, Larus' heading is certainly much more accurate.

I moved the if statement to OnSensorUpdate because high wind accuracy is desired. This is not guaranteed with OnCalculatedUpdate.

including relative wind vector

Including relative wind vector from external device like Larus, Anemoi, etc. If external wind is not available calculated wind from XCSoar/OpenSoar is displayed.

conversion of basic.track.Degrees() to integer removed

- conversion of basic.track.Degrees() to integer removed
- checked for the calculated wind, whether the external wind is definitely invalid